backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
Pillow==11.1.0
numpy==2.2.3
